feat: create requests validation

+++ src/main/java/br/com/finance/dtos/response/RevenueResponseDto.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 
 import br.com.finance.models.Revenue;
 
-public record RevenueResponseDto(Long id, String description, int value, LocalDate date, Long user_id) {
+public record RevenueResponseDto(Long id, String description, double value, LocalDate date, Long user_id) {
 
   public RevenueResponseDto(Revenue revenue){
     this(revenue.getId(), revenue.getDescription(), revenue.getValue(), revenue.getDate(), revenue.getUser().getId());
